feat: Remove Spring Security filter in user server
