+++ gateway/src/main/java/cm/twentysix/gateway/security/JwtAuthenticationFilter.java
@@ -11,8 +11,6 @@ import org.springframework.http.server.reactive.ServerHttpResponse;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.ReactiveSecurityContextHolder;
-import org.springframework.security.core.context.SecurityContext;
-import org.springframework.security.core.context.SecurityContextImpl;
 import org.springframework.stereotype.Component;
 import org.springframework.util.ObjectUtils;
 import org.springframework.util.StringUtils;
@@ -22,6 +20,7 @@ import org.springframework.web.server.WebFilterChain;
 import reactor.core.publisher.Mono;
 
 import java.util.Collections;
+import java.util.Optional;
 
 @RequiredArgsConstructor
 @Component
@@ -32,31 +31,51 @@ public class JwtAuthenticationFilter implements WebFilter {
 
     @Override
     public Mono<Void> filter(ServerWebExchange exchange, WebFilterChain chain) {
-        return getToken(exchange.getRequest())
-                .flatMap(accessToken -> processToken(exchange, chain, accessToken))
-                .switchIfEmpty(chain.filter(exchange))
-                .then();
-    }
-
-    private Mono<Void> processToken(ServerWebExchange exchange, WebFilterChain chain, String accessToken) {
-        ServerHttpRequest request = exchange.getRequest();
-        ServerHttpResponse response = exchange.getResponse();
-
-        try {
-            if (isValidAccessToken(accessToken)) {
-                return authorizeAndFilter(chain, exchange, accessToken);
+        Optional<String> maybeToken = getToken(exchange.getRequest());
+        if (maybeToken.isPresent()) {
+            String accessToken = maybeToken.get();
+            try {
+                if (StringUtils.hasText(accessToken) && jwtTokenManager.validate(accessToken)) {
+                    return authorize(accessToken, exchange, chain);
+                }
+            } catch (ExpiredJwtException e) {
+                return authorizeIfValidRefreshToken(exchange.getRequest(), exchange.getResponse(), exchange, chain);
             }
-        } catch (ExpiredJwtException e) {
-            return authorizeWithRefreshToken(request, response, chain, exchange);
         }
-
-        return chain.filter(exchange).contextWrite(ReactiveSecurityContextHolder.withSecurityContext(Mono.just(new SecurityContextImpl())));
+        return chain.filter(exchange);
     }
 
-    private Mono<Void> authorizeAndFilter(WebFilterChain chain, ServerWebExchange exchange, String accessToken) {
-        SecurityContext context = authorize(accessToken);
+    private Optional<String> getToken(ServerHttpRequest request) {
+        String rawToken = request.getHeaders().getFirst(HttpHeaders.AUTHORIZATION);
+        if (!ObjectUtils.isEmpty(rawToken) && rawToken.startsWith(TOKEN_PREFIX))
+            return Optional.of(rawToken.substring(TOKEN_PREFIX.length()));
+        return Optional.empty();
+    }
+
+    private Mono<Void> authorizeIfValidRefreshToken(ServerHttpRequest request, ServerHttpResponse response, ServerWebExchange exchange, WebFilterChain chain) {
+        Optional<String> maybeValidToken = CookieUtil.getCookieValue(request.getCookies(), "refreshToken")
+                .stream().findFirst()
+                .filter(jwtTokenManager::isValidRefreshToken);
+        if (maybeValidToken.isEmpty())
+            return chain.filter(exchange);
+
+        String refreshToken = maybeValidToken.get();
+
+        jwtTokenManager.deleteRefreshToken(refreshToken);
+        String userId = jwtTokenManager.parseId(refreshToken);
+        String userRole = jwtTokenManager.parseType(refreshToken);
+        String newAccessToken = jwtTokenManager.makeAccessToken(userId, userRole);
+        String newRefreshToken = jwtTokenManager.makeRefreshTokenAndSave(userId, userRole);
+
+        response.getHeaders().set(HttpHeaders.AUTHORIZATION, TOKEN_PREFIX + newAccessToken);
+        response.getHeaders().set(HttpHeaders.SET_COOKIE, CookieUtil.makeCookie("refreshToken", newRefreshToken));
+        return authorize(newAccessToken, exchange, chain);
+    }
+
+    private Mono<Void> authorize(String accessToken, ServerWebExchange exchange, WebFilterChain chain) {
         exchange = addUserIdToHeader(exchange, accessToken);
-        return chain.filter(exchange).contextWrite(ReactiveSecurityContextHolder.withSecurityContext(Mono.just(context)));
+        String userRole = jwtTokenManager.parseType(accessToken);
+        return chain.filter(exchange).contextWrite(ReactiveSecurityContextHolder.withAuthentication(new UsernamePasswordAuthenticationToken("user", null, Collections.singleton(new SimpleGrantedAuthority(userRole)))));
     }
 
     private ServerWebExchange addUserIdToHeader(ServerWebExchange exchange, String accessToken) {
@@ -67,51 +86,5 @@ public class JwtAuthenticationFilter implements WebFilter {
         return exchange.mutate().request(modifiedRequest).build();
     }
 
-    private boolean isValidAccessToken(String accessToken) {
-        return StringUtils.hasText(accessToken) && jwtTokenManager.validate(accessToken);
-    }
-
-    private Mono<Void> authorizeWithRefreshToken(ServerHttpRequest request, ServerHttpResponse response, WebFilterChain chain, ServerWebExchange exchange) {
-        return getValidRefreshToken(request)
-                .map(refreshToken -> renewTokensAndAuthorize(response, refreshToken))
-                .defaultIfEmpty(new SecurityContextImpl())
-                .flatMap(context -> chain.filter(exchange).contextWrite(ReactiveSecurityContextHolder.withSecurityContext(Mono.just(context))));
-    }
-
-    private Mono<String> getToken(ServerHttpRequest request) {
-        String rawToken = request.getHeaders().getFirst(HttpHeaders.AUTHORIZATION);
-        if (!ObjectUtils.isEmpty(rawToken) && rawToken.startsWith(TOKEN_PREFIX)) {
-            return Mono.just(rawToken.substring(TOKEN_PREFIX.length()));
-        }
-        return Mono.empty();
-    }
-
-    private Mono<String> getValidRefreshToken(ServerHttpRequest request) {
-        return Mono.justOrEmpty(CookieUtil.getCookieValue(request.getCookies(), "refreshToken")
-                .stream()
-                .findFirst()
-                .filter(jwtTokenManager::isValidRefreshToken));
-    }
-
-    private SecurityContext renewTokensAndAuthorize(ServerHttpResponse response, String refreshToken) {
-        jwtTokenManager.deleteRefreshToken(refreshToken);
-
-        String userId = jwtTokenManager.parseId(refreshToken);
-        String userRole = jwtTokenManager.parseType(refreshToken);
-        String newAccessToken = jwtTokenManager.makeAccessToken(userId, userRole);
-        String newRefreshToken = jwtTokenManager.makeRefreshTokenAndSave(userId, userRole);
-
-        response.getHeaders().set(HttpHeaders.AUTHORIZATION, TOKEN_PREFIX + newAccessToken);
-        response.getHeaders().set(HttpHeaders.SET_COOKIE, CookieUtil.makeCookie("refreshToken", newRefreshToken));
-
-        return authorize(newAccessToken);
-    }
-
-    private SecurityContext authorize(String accessToken) {
-        String userRole = jwtTokenManager.parseType(accessToken);
-        return new SecurityContextImpl(
-                new UsernamePasswordAuthenticationToken("user", null, Collections.singleton(new SimpleGrantedAuthority(userRole)))
-        );
-    }
 
 }
