+++ user/src/main/java/cm/twentysix/user/client/MailgunClient.java
@@ -1,6 +1,6 @@
 package cm.twentysix.user.client;
 
-import cm.twentysix.user.client.dto.SendMailForm;
+import cm.twentysix.user.dto.SendMailForm;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.cloud.openfeign.SpringQueryMap;
