+++ user/src/main/java/cm/twentysix/user/controller/EmailAuthController.java
@@ -1,7 +1,7 @@
 package cm.twentysix.user.controller;
 
-import cm.twentysix.user.controller.dto.SendAuthEmailForm;
-import cm.twentysix.user.controller.dto.SendAuthEmailResponse;
+import cm.twentysix.user.dto.SendAuthEmailForm;
+import cm.twentysix.user.dto.SendAuthEmailResponse;
 import cm.twentysix.user.service.EmailAuthService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
