+++ user/src/main/java/cm/twentysix/user/controller/LogInController.java
@@ -1,16 +1,13 @@
 package cm.twentysix.user.controller;
 
-import cm.twentysix.user.controller.dto.LogInForm;
-import cm.twentysix.user.controller.dto.TokenResponse;
+import cm.twentysix.user.dto.LogInForm;
+import cm.twentysix.user.dto.TokenResponse;
 import cm.twentysix.user.service.LogInService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
