+++ user/src/main/java/cm/twentysix/user/controller/SignUpController.java
@@ -1,6 +1,6 @@
 package cm.twentysix.user.controller;
 
-import cm.twentysix.user.controller.dto.SignUpForm;
+import cm.twentysix.user.dto.SignUpForm;
 import cm.twentysix.user.service.SignUpService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
