+++ user/src/main/java/cm/twentysix/user/dto/LogOutType.java
@@ -1,4 +1,4 @@
-package cm.twentysix.user.controller.dto;
+package cm.twentysix.user.dto;
 
 public enum LogOutType {
     ALL, SINGLE
