+++ user/src/main/java/cm/twentysix/user/dto/SendAuthEmailForm.java
@@ -1,4 +1,4 @@
-package cm.twentysix.user.controller.dto;
+package cm.twentysix.user.dto;
 
 import jakarta.validation.constraints.Pattern;
