+++ user/src/main/java/cm/twentysix/user/dto/SendAuthEmailResponse.java
@@ -1,4 +1,4 @@
-package cm.twentysix.user.controller.dto;
+package cm.twentysix.user.dto;
 
 import lombok.Builder;
