--- user/src/main/java/cm/twentysix/user/security/UserPrincipal.java
@@ -1,51 +1,0 @@
-package cm.twentysix.user.security;
-
-import lombok.Getter;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-
-@Getter
-public class UserPrincipal implements UserDetails {
-    private final Long userId;
-
-    public UserPrincipal(Long userId) {
-        this.userId = userId;
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return null;
-    }
-
-    @Override
-    public String getPassword() {
-        return String.valueOf(userId);
-    }
-
-    @Override
-    public String getUsername() {
-        return String.valueOf(userId);
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
