--- user/src/main/java/cm/twentysix/user/security/UserPrincipalService.java
@@ -1,16 +1,0 @@
-package cm.twentysix.user.security;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class UserPrincipalService implements UserDetailsService {
-    @Override
-    public UserDetails loadUserByUsername(String userId) throws UsernameNotFoundException {
-        return new UserPrincipal(Long.parseLong(userId));
-    }
-}
